perf: 로딩페이지 성능 개선
