+++ src/app/[locale]/auth/loading/RedirectAfterDelay.jsx
@@ -1,0 +1,18 @@
+"use client";
+
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+
+export default function RedirectAfterDelay() {
+  const router = useRouter();
+
+  useEffect(() => {
+    const id = setTimeout(() => {
+      router.replace("/auth/success");
+    }, 3000);
+
+    return () => clearTimeout(id);
+  }, [router]);
+
+  return null;
+}
