+++ src/app/[locale]/main/components/MoneyExchange/RealTimeExchange/RealTimeExchange.jsx
@@ -20,6 +20,8 @@ export default function RealTimeExchange() {
         const usdData = data.find((item) => item.cur_unit === "USD");
         setUsd(usdData);
         setIsLoading(false);
+
+        console.log("환율 조회 데이터: ", data);
       })
       .catch((error) => {
         console.error("환율 데이터 조회 실패:", error);
